fix broken modules from merge
